fix: resolved comments and some refactoring
