+++ frontend/src/container/NewWidget/RightContainer/Threshold/Threshold.styles.scss
@@ -300,7 +300,7 @@
 
 		.invalid-unit {
 			color: var(--bg-vanilla-400);
-			font-family: 'Space Mono';
+			font-family: 'Giest Mono';
 			font-size: 11px;
 			font-style: normal;
 			font-weight: 400;
